[refactor] wrapper Type 테스트 케이스를 따로 옮김
