+++ src/test/java/com/inhabas/api/domain/BoardMappingTest.java
@@ -1,0 +1,9 @@
+package com.inhabas.api.domain;
+
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+@DataJpaTest
+public class BoardMappingTest {
+
+
+}
