+++ src/test/java/com/inhabas/api/domain/wrapper/FilenameTest.java
@@ -1,0 +1,36 @@
+package com.inhabas.api.domain.wrapper;
+
+import com.inhabas.api.domain.file.type.wrapper.FileName;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class FilenameTest {
+
+    @DisplayName("FileName 타입에 파일 이름을 저장한다.")
+    @Test
+    public void FileName_is_OK() {
+        //when
+        FileName fileName = new FileName("filename_test_file.txt");
+
+        //then
+        assertThat(fileName.getValue()).isEqualTo("filename_test_file.txt");
+        assertThat(fileName.getExtension()).isEqualTo("txt");
+        assertThat(fileName.getNameWithoutExtension()).isEqualTo("filename_test_file");
+    }
+
+    @DisplayName("FileName 타입에 너무 긴 파일 이름 저장을 시도한다.")
+    @Test
+    public void FileName_is_too_long() {
+        //given
+        String tooLongFileName = "a".repeat(500) + "txt";
+
+        //when
+        assertThrows(
+                IllegalArgumentException.class,
+                () -> { FileName fileName = new FileName(tooLongFileName); }
+        );
+    }
+}
