+++ src/test/java/com/inhabas/api/domain/wrapper/FilepathTest.java
@@ -1,0 +1,39 @@
+package com.inhabas.api.domain.wrapper;
+
+import com.inhabas.api.domain.file.type.wrapper.FilePath;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class FilepathTest {
+
+    @DisplayName("FilePath 타입에 파일 경로를 저장한다.")
+    @Test
+    public void FilePath_is_OK() {
+        //given
+        String path = "/file/board/";
+        String filename = "file.txt";
+
+        //when
+        FilePath filePath = new FilePath(path + filename);
+
+        //then
+        assertThat(filePath.getValue()).isEqualTo("/file/board/file.txt");
+    }
+
+    @DisplayName("FilePath 타입에 너무 긴 경로 저장을 시도한다.")
+    @Test
+    public void FilePath_is_too_long() {
+        //given
+        String path = "/" + "file".repeat(300);
+        String tooLongFilePath = path + "file.txt";
+
+        //when
+        assertThrows(
+                IllegalArgumentException.class,
+                ()->{FilePath filePath = new FilePath(tooLongFilePath);}
+        );
+    }
+}
