+++ src/test/java/com/inhabas/api/domain/wrapper/IntroduceTest.java
@@ -1,0 +1,37 @@
+package com.inhabas.api.domain.wrapper;
+
+import com.inhabas.api.domain.member.type.wrapper.Introduce;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class IntroduceTest {
+
+    @DisplayName("자기소개란 작성")
+    @Test
+    public void Introduce_is_OK() {
+        //given
+        String introduceString = "아이엠 그라운드 자기소개 하기.";
+
+        //when
+        Introduce introduce = new Introduce(introduceString);
+
+        //then
+        assertThat(introduce.getValue()).isEqualTo("아이엠 그라운드 자기소개 하기.");
+    }
+
+    @DisplayName("자기소개가 너무 길다. 300자 이상.")
+    @Test
+    public void Introduce_is_too_long() {
+        //given
+        String introduceString = "지금이문장은10자임".repeat(30); // 300자
+
+        //then
+        assertThrows(
+                IllegalArgumentException.class,
+                ()-> new Introduce(introduceString)
+        );
+    }
+}
